マークをスキャンする処理に入れ替え
